Allow for order fulfillment
What?
=====

This introduces the core of the ordering and fulfillment logic via
introduction of a few core domain objects: Order, OrderLineItem, and
Address. With these structures in place, order fulfillment/shipment is
now tied to a particular order, and an order can be inferred as
fulfilled or not due to one-shot fulfillment (an order cannot be
partially fulfilled).

With this in place, alongside simple UI fulfillment, employees can
fulfill orders where full quantities are covered with a single click. On
the employee dashboard, the newest orders are displayed, allowing for
those recent orders to be fulfilled.

This introduces some fairly complex logic to determine which orders are
able to be fulfilled via Arel. This was originally written with raw SQL,
but with the introduction of lookup by order ID, further abstractions
were necessary.

Finally, this updates the dev_seed task to fulfill as many orders as
possible, helping to simulate real-world status updates and order flow.