Introduce a `DevSimuation` class to simulate real-time ordering
What?
=====

This introduces a new class, `DevSimuation`, which manages simulating
inbound orders with a default configuration around opportunity
frequency, likelihood of an order occurring, and likelihood of it being
a repeat customer.

This also includes some tests around configuration settings to ensure
the configuration works and the simulation will run correctly.